move color to its own file, add basic (unused) implementations of Spritesheet and Sprite
