+++ src/util/quantize.ts
@@ -1,7 +1,7 @@
-import Bitmap4 from "../models/Bitmap4";
-import Palette from "../models/Palette";
-import { Color } from "./types";
 import Bitmap from "../models/Bitmap";
+import Bitmap4 from "../models/Bitmap4";
+import Color from "../models/Color";
+import Palette from "../models/Palette";
 
 const BLACK: Color = new Color(0, 0, 0, 1);
